# pb/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/tally/counter": {
      "get": {
        "operationId": "GetCounter",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbGetCounterResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startEpochMinute",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "endEpochMinute",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "QueryCounterService"
        ]
      }
    },
    "/v1/tally/counter/record": {
      "post": {
        "operationId": "RecordCounter",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbRecordCounterResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbRecordCounterRequest"
            }
          }
        ],
        "tags": [
          "RecordCounterService"
        ]
      }
    }
  },
  "definitions": {
    "pbCounterValues": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "sum": {
          "type": "number",
          "format": "float"
        },
        "min": {
          "type": "number",
          "format": "float"
        },
        "max": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "pbGetCounterRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "source": {
          "type": "string"
        },
        "startEpochMinute": {
          "type": "integer",
          "format": "int32"
        },
        "endEpochMinute": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "pbGetCounterResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean"
        },
        "error": {
          "type": "string"
        },
        "values": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/pbCounterValues"
          }
        }
      }
    },
    "pbRecordCounterRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "source": {
          "type": "string"
        },
        "epochMinute": {
          "type": "integer",
          "format": "int32"
        },
        "values": {
          "$ref": "#/definitions/pbCounterValues"
        }
      }
    },
    "pbRecordCounterResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean"
        },
        "error": {
          "type": "string"
        }
      }
    }
  }
}